+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - fix seerr sonarr activeDirectory default resolving host config ([#283](https://github.com/kiriwalawren/nixflix/pull/283))
 - fix \*arr rootfolders oneshot racing app startup ([#286](https://github.com/kiriwalawren/nixflix/pull/286))
+- fix \*arr `config.hostConfig` auth/port/urlBase values being ignored in favor of hardcoded `settings` env vars ([#289](https://github.com/kiriwalawren/nixflix/issues/289))
 
 ## [3.0.0] - 2026-07-12